+++ golded3/gehdrs.cpp
@@ -118,8 +118,8 @@ void DispHeader(GMsg* msg, bool prn, FILE* fp, int width) {
   // Generate message number and reply links string
   char* ptr = buf1;
   int list_max = msg->link.list_max();
-  ulong* replies = (ulong*)throw_calloc(list_max+1, sizeof(ulong));
-  ulong replyto, replynext;
+  uint32_t* replies = (uint32_t*)throw_calloc(list_max+1, sizeof(uint32_t));
+  uint32_t replyto, replynext;
   if(CFG->switches.get(disprealmsgno)) {
     ptr += sprintf(ptr, " %-5.5s: #%lu [%u]", LNG->Msg, msg->msgno, AA->Msgn.Count()+(msg->attr.nwm() ? 1 : 0));
     replyto = msg->link.to();
